Запустил всё в Docker, небольшие изменения в для корректной работы Liquibase

+++ registration-service/registration-service-api/src/main/java/com/targetmonkey/registrationserviceapi/resource/v1/AdCleintRegistrationV1.java
@@ -16,12 +16,12 @@ public interface AdCleintRegistrationV1 {
     List<AdCustomerRegistrationDTO> getAll();
 
     @GetMapping("customer/{id}")
-    AdCustomerRegistrationDTO getToId(@PathVariable long id);
+    AdCustomerRegistrationDTO getToId(@PathVariable int id);
 
     @PutMapping("edit/{id}")
-    String editAdCustomer(@PathVariable long id,
+    String editAdCustomer(@PathVariable int id,
                           @RequestBody AdCustomerRegistrationDTO adCustomerRegistrationDTO);
 
     @DeleteMapping("delete/{id}")
-    String deleteAdCustomer(@PathVariable long id);
+    String deleteAdCustomer(@PathVariable int id);
 }
