+++ registration-service/registration-service-impl/src/main/java/com/targetmonkey/registrationserviceimpl/repository/interfaces/AdCompanyRegRepository.java
@@ -3,5 +3,5 @@ package com.targetmonkey.registrationserviceimpl.repository.interfaces;
 import com.targetmonkey.registrationserviceimpl.entity.AdCompanyRegistrationJpa;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdCompanyRegRepository extends JpaRepository<AdCompanyRegistrationJpa, Long> {
+public interface AdCompanyRegRepository extends JpaRepository<AdCompanyRegistrationJpa, Integer> {
 }
